Refactor: Category DTO 분리
