+++ src/main/java/com/newzet/api/category/controller/dto/CategoryListResponse.java
@@ -9,15 +9,4 @@ public record CategoryListResponse(
 	public static CategoryListResponse create(List<CategoryResponse> categoryList) {
 		return new CategoryListResponse(categoryList);
 	}
-
-	public record CategoryResponse(
-		String id,
-		String name,
-		String imageUrl,
-		String emoji
-	) {
-		public static CategoryResponse create(UUID id, String name, String imageUrl, String emoji) {
-			return new CategoryResponse(id.toString(), name, imageUrl, emoji);
-		}
-	}
 }
